requirements.txt
flask==3.0.0
numpy==1.24.3
Pillow==10.1.0
onnxruntime==1.16.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
